QoL Changes

+++ Client/server_processes/deleteJsonFile.js
@@ -1,6 +1,5 @@
 /**
- * @name deleteJsonFile.js
- * @description Deletes a file.
+ * Deletes a file.
  *
  * @param file
  * @returns {*}
@@ -8,15 +7,15 @@
 module.exports = function (file) {
     let fs = require('fs');
 
-    let json;
-
     try {
-        fs.rmSync(file);
+        if (fs.existsSync(file)) {
+            fs.rmSync(file);
+        }
     } catch(error) {
+        const {dialog} = require('electron');
+        dialog.showErrorBox('Vault Error (3)',error.toString()+" - More help: bit.ly/vaulterrors");
         console.log('[VAULT] Failed to read JSON file "'+file+'"');
         console.log('[VAULT] Error: ' + error);
         console.log('[VAULT] Please check "'+file+'" is readable and try again.');
     }
-
-    return json;
 }
